port over series logic to plot item
