+++ sinks/dashboard/series.py
@@ -8,7 +8,7 @@ class Subject:
     """
     def __init__ (self):
         self.observers = []
-        
+
     def add_observer(self, observer):
         """
         An observer is a dashboard item that cares
@@ -34,20 +34,7 @@ class Series(Subject):
     def __init__(self, name, time_rollover=False):
         super().__init__()
         self.name = name
-
-        self.size = config.GRAPH_RESOLUTION * config.GRAPH_DURATION
-        self.last = 0
-        self.times = np.zeros(self.size)
-        self.points = np.zeros(self.size)
-        self.sum = 0  # sum of series
-        # "size" of running average
-        self.avgSize = config.RUNNING_AVG_DURATION * config.GRAPH_RESOLUTION
-        self.desc = None
-        self.time_rollover = time_rollover
-        self.time_offset = 0
-
         self.callback = None
-
 
     def add(self, time, point, desc=None):
         """
@@ -55,36 +42,11 @@ class Series(Subject):
         """
         if desc is not None:
             self.desc = desc
-
-        time += self.time_offset
-        if self.time_rollover:
-            if time < self.times[-1]:  # if we've wrapped around
-                self.time_offset += self.times[-1]  # increase the amount we need to add
-
-        # time should be passed as seconds, GRAPH_RESOLUTION is points per second
-        if time - self.last < 1 / config.GRAPH_RESOLUTION:
-            return
-
-        if self.last == 0:  # is this the first point we're plotting?
-            self.times.fill(time)  # prevent a rogue datapoint at (0, 0)
-            self.points.fill(point)
-            self.sum = self.avgSize * point
-
-        self.last += 1 / config.GRAPH_RESOLUTION
-
-        self.sum -= self.points[self.size - self.avgSize]
-        self.sum += point
-
-        # add the new datapoint to the end of each array, shuffle everything else back
-        self.times[:-1] = self.times[1:]
-        self.times[-1] = time
-        self.points[:-1] = self.points[1:]
-        self.points[-1] = point
         
+        self.time = time
+        self.point = point
         self.notify_observers()
 
-    def get_running_avg(self):
-        return self.sum / self.avgSize
 
 
 class CanMsgSeries(Subject):
